revno: 10268 message:   report conflict if user is not allowed to update object
committer: Morten Olav Hansen <mortenoh@gmail.com>
branch nick: dhis2
timestamp: Sat 2013-03-16 21:22:03 +0700

+++ dhis-2/dhis-dxf2/src/main/java/org/hisp/dhis/dxf2/metadata/importers/DefaultIdentifiableObjectImporter.java
@@ -36,6 +36,7 @@ import org.hisp.dhis.attribute.AttributeValue;
 import org.hisp.dhis.common.BaseIdentifiableObject;
 import org.hisp.dhis.common.IdentifiableObject;
 import org.hisp.dhis.common.NameableObject;
+import org.hisp.dhis.common.SharingUtils;
 import org.hisp.dhis.dataelement.DataElementOperand;
 import org.hisp.dhis.dataelement.DataElementOperandService;
 import org.hisp.dhis.dataentryform.DataEntryForm;
@@ -58,6 +59,7 @@ import org.hisp.dhis.period.PeriodType;
 import org.hisp.dhis.system.util.CollectionUtils;
 import org.hisp.dhis.system.util.ReflectionUtils;
 import org.hisp.dhis.system.util.functional.Function1;
+import org.hisp.dhis.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.lang.reflect.Field;
@@ -417,12 +419,21 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
     /**
      * Update idObject from old => new.
      *
+     * @param user
      * @param object          Object to import
      * @param persistedObject The current version of the idObject
      * @return An ImportConflict instance if there was a conflict, otherwise null
      */
-    protected boolean updateObject( T object, T persistedObject )
+    protected boolean updateObject( User user, T object, T persistedObject )
     {
+        if ( !SharingUtils.canUpdate( user, persistedObject ) )
+        {
+            summaryType.getImportConflicts().add(
+                new ImportConflict( ImportUtils.getDisplayName( object ), "You do not have update access to object." ) );
+
+            return false;
+        }
+
         NonIdentifiableObjects nonIdentifiableObjects = new NonIdentifiableObjects();
         nonIdentifiableObjects.extract( object );
         nonIdentifiableObjects.delete( persistedObject );
@@ -439,11 +450,6 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
 
         log.debug( "Starting update of object " + ImportUtils.getDisplayName( persistedObject ) + " (" + persistedObject.getClass()
             .getSimpleName() + ")" );
-
-        if ( persistedObject.getName().contains( "java" ) )
-        {
-            System.err.println( "clazz: " + persistedObject.getClass().getName() + ", persistedObject: " + persistedObject );
-        }
 
         objectBridge.updateObject( persistedObject );
 
@@ -475,7 +481,7 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
     //-------------------------------------------------------------------------------------------------------
 
     @Override
-    public ImportTypeSummary importObjects( List<T> objects, ImportOptions options )
+    public ImportTypeSummary importObjects( User user, List<T> objects, ImportOptions options )
     {
         this.options = options;
         this.summaryType = new ImportTypeSummary( importerClass.getSimpleName() );
@@ -490,7 +496,7 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
         for ( T object : objects )
         {
             ObjectHandlerUtils.preObjectHandlers( object, objectHandlers );
-            importObjectLocal( object );
+            importObjectLocal( user, object );
             ObjectHandlerUtils.postObjectHandlers( object, objectHandlers );
         }
 
@@ -500,13 +506,13 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
     }
 
     @Override
-    public ImportTypeSummary importObject( T object, ImportOptions options )
+    public ImportTypeSummary importObject( User user, T object, ImportOptions options )
     {
         this.options = options;
         this.summaryType = new ImportTypeSummary( importerClass.getSimpleName() );
 
         ObjectHandlerUtils.preObjectHandlers( object, objectHandlers );
-        importObjectLocal( object );
+        importObjectLocal( user, object );
         ObjectHandlerUtils.postObjectHandlers( object, objectHandlers );
 
         return summaryType;
@@ -522,11 +528,11 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
     // Helpers
     //-------------------------------------------------------------------------------------------------------
 
-    private void importObjectLocal( T object )
+    private void importObjectLocal( User user, T object )
     {
         if ( validateIdentifiableObject( object ) )
         {
-            startImport( object );
+            startImport( user, object );
         }
         else
         {
@@ -534,7 +540,7 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
         }
     }
 
-    private void startImport( T object )
+    private void startImport( User user, T object )
     {
         T oldObject = objectBridge.getObject( object );
 
@@ -547,7 +553,7 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
         }
         else if ( ImportStrategy.UPDATES.equals( options.getImportStrategy() ) )
         {
-            if ( updateObject( object, oldObject ) )
+            if ( updateObject( user, object, oldObject ) )
             {
                 summaryType.incrementUpdated();
             }
@@ -556,7 +562,7 @@ public class DefaultIdentifiableObjectImporter<T extends BaseIdentifiableObject>
         {
             if ( oldObject != null )
             {
-                if ( updateObject( object, oldObject ) )
+                if ( updateObject( user, object, oldObject ) )
                 {
                     summaryType.incrementUpdated();
                 }
